requirements.txt
flask==2.0.1
spotipy==2.19.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.86.0
python-dotenv==1.0.0
requests==2.26.0
